Split `non_local_definitions` lint tests in separate test files
