+++ tests/ui/lint/non-local-defs/auxiliary/non_local_macro.rs
