+++ tests/ui/lint/non-local-defs/cargo-update.rs
@@ -1,0 +1,20 @@
+//@ check-pass
+//@ edition:2021
+//@ aux-build:non_local_macro.rs
+//
+// To suggest any Cargo specific help/note rustc wants
+// the `CARGO_CRATE_NAME` env to be set, so we set it
+//@ rustc-env:CARGO_CRATE_NAME=non_local_def
+//
+// and since we specifically want to check the presence
+// of the `cargo update` suggestion we assert it here.
+//@ error-pattern: `cargo update -p non_local_macro`
+
+extern crate non_local_macro;
+
+struct LocalStruct;
+
+non_local_macro::non_local_impl!(LocalStruct);
+//~^ WARN non-local `impl` definition
+
+fn main() {}
