+++ tests/ui/lint/non-local-defs/inside-macro_rules.rs
@@ -1,0 +1,17 @@
+//@ check-pass
+//@ edition:2021
+
+macro_rules! m {
+    () => {
+        trait MacroTrait {}
+        struct OutsideStruct;
+        fn my_func() {
+            impl MacroTrait for OutsideStruct {}
+            //~^ WARN non-local `impl` definition
+        }
+    }
+}
+
+m!();
+
+fn main() {}
